area: extract windows related keymaps to separate file to reuse in the entry
